einet: Uncertainty and causal emergence in complex networks.

for calculating effective information in networks. This can then be used to search for macroscale representations of a network such that the coarse grained representation has more effective information than the microscale, a phenomenon known as causal emergence.

See also

Author

Maintainer: Travis Byrum tbyrum@terpmail.umd.edu

Authors:

Anshuman Swain answain@terpmail.umd.edu

Brennan Klein klein.br@northeastern.edu

William Fagan bfagan@umd.edu